+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/psi/impl/xpath/XPathElementTestPsiImpl.kt
@@ -17,8 +17,11 @@ package uk.co.reecedunn.intellij.plugin.xpath.psi.impl.xpath
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.core.data.CacheableProperty
 import uk.co.reecedunn.intellij.plugin.core.sequences.children
+import uk.co.reecedunn.intellij.plugin.intellij.lang.*
+import uk.co.reecedunn.intellij.plugin.intellij.resources.XPathBundle
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathElementTest
 import uk.co.reecedunn.intellij.plugin.xdm.functions.op.op_qname_presentation
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmElementNode
@@ -26,14 +29,28 @@ import uk.co.reecedunn.intellij.plugin.xdm.types.XdmItemType
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmSequenceType
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathTypeName
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathWildcard
 
-class XPathElementTestPsiImpl(node: ASTNode) : ASTWrapperPsiElement(node), XPathElementTest, XdmItemType {
+class XPathElementTestPsiImpl(node: ASTNode) :
+    ASTWrapperPsiElement(node), VersionConformance, VersionConformanceName, XPathElementTest, XdmItemType {
     // region ASTDelegatePsiElement
 
     override fun subtreeChanged() {
         super.subtreeChanged()
         cachedTypeName.invalidate()
     }
+
+    // endregion
+    // region VersionConformance
+
+    override val requiresConformance: List<Version>
+        get() = if (conformanceElement is XPathWildcard) listOf(Saxon.VERSION_10_0) else listOf()
+
+    override val conformanceElement: PsiElement
+        get() = children().filterIsInstance<XPathWildcard>().firstOrNull() ?: firstChild
+
+    override val conformanceName: String?
+        get() = XPathBundle.message("construct.wildcard-element-test")
 
     // endregion
     // region XPathElementTest